+++ daemon/x52d_mouse_test.c
@@ -44,75 +44,61 @@ static void test_mouse_thread_disabled(void **state)
     x52d_cfg_set_Mouse_Enabled(false);
 }
 
+/* The following tests are dependent on the values in x52d_mouse.c */
 static void test_mouse_speed_negative(void **state)
 {
     (void)state;
-    int orig_mouse_delay = mouse_delay;
-    int orig_mouse_mult = mouse_mult;
-
     x52d_cfg_set_Mouse_Speed(-1);
-    assert_int_equal(mouse_delay, orig_mouse_delay);
-    assert_int_equal(mouse_mult, orig_mouse_mult);
+    assert_int_equal(mouse_sensitivity, 14);
 }
 
-/* The following tests are dependent on the values in x52d_mouse.c */
 static void test_mouse_speed_0(void **state)
 {
     (void)state;
     x52d_cfg_set_Mouse_Speed(0);
-    assert_int_equal(mouse_delay, 70000);
-    assert_int_equal(mouse_mult, 4);
+    assert_int_equal(mouse_sensitivity, 14);
 }
 
 static void test_mouse_speed_mid_base(void **state)
 {
     (void)state;
     x52d_cfg_set_Mouse_Speed(6);
-    assert_int_equal(mouse_delay, 40000);
-    assert_int_equal(mouse_mult, 4);
+    assert_int_equal(mouse_sensitivity, 25);
 }
 
 static void test_mouse_speed_max_base(void **state)
 {
     (void)state;
     x52d_cfg_set_Mouse_Speed(12);
-    assert_int_equal(mouse_delay, 10000);
-    assert_int_equal(mouse_mult, 4);
+    assert_int_equal(mouse_sensitivity, 100);
 }
 
 static void test_mouse_speed_min_hyper(void **state)
 {
     (void)state;
     x52d_cfg_set_Mouse_Speed(13);
-    assert_int_equal(mouse_delay, 10000);
-    assert_int_equal(mouse_mult, 5);
+    assert_int_equal(mouse_sensitivity, 125);
 }
 
 static void test_mouse_speed_mid_hyper(void **state)
 {
     (void)state;
     x52d_cfg_set_Mouse_Speed(22);
-    assert_int_equal(mouse_delay, 10000);
-    assert_int_equal(mouse_mult, 14);
+    assert_int_equal(mouse_sensitivity, 350);
 }
 
 static void test_mouse_speed_max_hyper(void **state)
 {
     (void)state;
     x52d_cfg_set_Mouse_Speed(32);
-    assert_int_equal(mouse_delay, 10000);
-    assert_int_equal(mouse_mult, 24);
+    assert_int_equal(mouse_sensitivity, 500);
 }
 
 static void test_mouse_speed_above_max(void **state)
 {
-    int orig_mouse_delay = mouse_delay;
-    int orig_mouse_mult = mouse_mult;
     (void)state;
-
     x52d_cfg_set_Mouse_Speed(33);
-    assert_int_equal(mouse_delay, orig_mouse_delay);
-    assert_int_equal(mouse_mult, orig_mouse_mult);
+    assert_int_equal(mouse_sensitivity, 500);
 }
 
 static void test_mouse_reverse_scroll_enabled(void **state)